\name{.calcPortfAttr}
\alias{.calcPortfAttr}
\title{calculate summary attributes for a portfolio...}
\usage{.calcPortfAttr(Portfolio, Attribute, Dates, Symbols)}
\description{calculate summary attributes for a portfolio}
\details{supported values are:
Gross.Trading.PL, Txn.Fee, Net.Trading.PL, Realized.PL, Unrealized.PL, Gross.Value, Long.Value, Short.Value}
\arguments{\item{Portfolio}{portfolio object}
\item{Attribute}{attribute to calculate}
\item{Dates}{date subset}
\item{Symbols}{symbols to calculate for}}
