\name{.getPortfAcct}
\alias{.getPortfAcct}
\title{get a protfolio in an account...}
\usage{.getPortfAcct(Account, Portfolio, Dates)}
\description{get a protfolio in an account}
\arguments{\item{Account}{account string}
\item{Portfolio}{portfolio string}
\item{Dates}{date subset as an xts style ISO 8601 string}}
