\name{.initSummary}
\alias{.initSummary}
\title{initialize the summary table used in portfolio and account lists...}
\usage{.initSummary(initDate="1950-01-01")}
\description{initialize the summary table used in portfolio and account lists}
\arguments{\item{initDate}{date prior to the first close price given, used to contain initial account equity and initial position}}
