% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ABC}
\alias{ABC}
\title{Prices of random security ABC for a full trading day.}
\format{An xts object with 163,936 rows and 2 variables:
\describe{
  \item{TradePrice}{price in cents}
  \item{TradeVolume}{traded volume}
  ...
}}
\source{
\url{}
}
\usage{
ABC
}
\description{
To be used for among other things testing the TCA models being
added to blotter as part of GSoC 2019 by Vito Lestingi.
}
\keyword{datasets}
