% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addPortfInstr.R
\name{addPortfInstr}
\alias{addPortfInstr}
\title{add an instrument to a portfolio}
\usage{
addPortfInstr(Portfolio, symbols, ...)
}
\arguments{
\item{Portfolio}{portfolio identifier string}

\item{symbols}{character vector of symbols to add to the portfolio}

\item{\dots}{any other passthru parameters}
}
\description{
thanks to WolfGang Wu for making this function more usable
}
