\name{amzn}
\docType{data}
\alias{amzn}
\alias{amzn.trades}
\title{Hypothetical Intra-Day AMZN Trades}
\description{
Two xts objects containing 1-min OHLCV data for one hour of AMZN market activity, 
and object 'amzn.trades' containing hypothetical transactions over that period.
}
\usage{managers}
\keyword{datasets}
\keyword{ ts } 
