% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndexAdjustedCost.R
\name{ArrivalCost}
\alias{ArrivalCost}
\title{Arrival Cost}
\usage{
ArrivalCost(Portfolio, Symbol, side = "Buy")
}
\arguments{
\item{Portfolio}{A portfolio name that points to a portfolio object structured with initPortf()}

\item{Symbol}{An instrument identifier for a symbol included in the portfolio, e.g., "IBM"}

\item{side}{string identifying either "Buy" or "Sell"}
}
\value{
A numeric value in basis points
}
\description{
The Arrival Cost (or Trading Cost) is the difference between the
average execution price and the price of the asset at the time the
order was entered into the market, ie. the Arrival Price.
}
\details{
\deqn{ArrivalCost_bp} = Side \dot \frac{P_avg - P_0 }{P_0} \dot {10^4}_bp
\deqn{{p^s} = Pr( |r| > t-ratio)}{p^s = Pr(|r|>t-ratio)}
}
\examples{
ArrivalCost("testport", "IBM", side = "Buy")

}
