% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchTradePerf.R
\name{benchTradePerf}
\alias{benchTradePerf}
\title{Trade Execution Performance Benchmarks}
\usage{
benchTradePerf(Portfolio, Symbol, side = 1, benchmark = c("MktBench",
  "VWAP", "PWP", "RPM"), type = c("Txns", "Mkt"), MktData, POV = NULL,
  verbose = FALSE)
}
\arguments{
\item{Portfolio}{A portfolio name that points to a portfolio object structured with initPortf()}

\item{Symbol}{A string identifying the traded symbol to benchmark}

\item{side}{A numeric value, that indicates the side of the trade. Either 1 or -1, \code{side = 1} (default) means "Buy" and \code{side = -1} is "Sell"}

\item{benchmark}{A string or vector of strings providing the one or several of the 'MktBench', 'VWAP', 'PWP' or 'RPM' benchmark metrics. Default is #TODO}

\item{type}{A string specifying the type of the benchmark used. Relevant only for \code{benchmark = 'MktBench'} and \code{benchmark = 'VWAP'}
In the former case it only changes the output format, while in the latter it also specifies the VWAP to be used. If \code{benchmark = 'VWAP'}, default is \code{type = 'Txns'}}

\item{MktData}{An xts object containing a 'MktPrice' and 'MktVolmn' required columns}

\item{POV}{A numeric value between 0 and 1, specifying the POV rate}

\item{verbose}{A logical value. It allows a RPM qualitative score appended. Default is \code{FALSE}}
}
\value{
A list whose elements correspond in number to the length of the \code{benchmark} provided.
Each element is a \code{data.frame} that can be one of the ones described below.

For \code{benchmark = 'MktBench'} it contains:
\describe{
     \item{\code{Symbol}: }{A string identifying the traded symbol to benchmark}
     \item{\code{Side}: }{The \code{side} of the trades, as "Buy" or "Sell"}
     \item{\code{Avg.Exec.Price}: }{Symbol transactions average execution price}
     \item{\code{MktBench.*}: }{The benchmark and an arbitrary \code{type} provided as input (e.g. 'Open', 'Close')}
     \item{\code{Performance}: }{The \emph{Benchmark PnL} performance, in bps}
}

For \code{benchmark = 'VWAP'} it contains:
\describe{
     \item{\code{Symbol}: }{A string identifying the traded symbol to benchmark}
     \item{\code{Side}: }{The \code{side} of the trades, as "Buy" or "Sell"}
     \item{\code{Avg.Exec.Price}: }{Symbol transactions average execution price}
     \item{\code{VWAP.*}: }{The benchmark and depending on \code{type} parameter either 'Txns' 'Mkt'}
     \item{\code{Performance}: }{The \emph{VWAP PnL} metric, in bps}
}

For \code{benchmark = 'PWP'} it contains:
\describe{
     \item{\code{Symbol}: }{A string identifying the traded symbol to benchmark}
     \item{\code{Side}: }{The \code{side} of the trades, as "Buy" or "Sell"}
     \item{\code{Avg.Exec.Price}: }{Symbol transactions average execution price}
     \item{\code{POV}: }{The POV rate of the order}
     \item{\code{PWP.Shares}: }{The ratio between the total unit traded and the POV rate}
     \item{\code{PWP.Price}: }{Volume weighted price of the first \code{PWP.Shares} traded}
     \item{\code{Performance}: }{The \emph{PWP PnL} metric, in bps}
}

For \code{benchmark = 'RPM'} it contains:
\describe{
     \item{\code{Symbol}: }{A string identifying the traded symbol to benchmark}
     \item{\code{Side}: }{The \code{side} of the trades, as "Buy" or "Sell".}
     \item{\code{Avg.Exec.Price}: }{Symbol transactions average execution price}
     \item{\code{t.Mkt.Volmn}: }{Total market volume over the order timespan}
     \item{\code{t.Fav.Volmn}: }{Total market volume over the order timespan for which the average execution price of a 'Buy' ('Sell') order was lower (greater) than market prices}
     \item{\code{t.Unfav.Volmn}: }{The opposite of \code{t.Fav.Volmn}}
     \item{\code{RPM}: }{The \emph{relative performance measure}. Decimal in the 0 to 1 range.}
     \item{\code{Quality}: }{A qualitiative RPM score over quintiles, bottom-up one of 'Poor', 'Fair', 'Average', 'Good', 'Excellent'. Present if \code{verbose = TRUE}}
}
}
\description{
This function gathers different benchmarking methods used to evaluate the 
average execution price \eqn{\bar{P}} of a given trading strategy.
The \eqn{\bar{P}} is compared against a number of benchmark metrics, in order
to assess its performance in terms of profit or loss relative to a given benchmark.
These benchmarks are not mutually exclusive, each of them provides different insights
and may have shortcomings. They can be used in conjuction to account for this aspects.
}
\details{
The performance is quantified by means of a \emph{Profit and Loss (PnL) metric}. 
A positive PnL metric indicates that the trading strategy outperformed a chosen 
benchmark on average, vice versa negative values register an underperformance.

By and large, PnL metrics are computed as:

#TODO

where \eqn{P_{avg}} is the average execution price and \eqn{P_{B}} is a given 
benchmark price that can be a single next/present open/close price.
It is worth stressing that they are expressed in basis points (bps) units.

A widely used one is the \emph{Volume Weighted Average Price (VWAP) benchamark}.
Where the VWAP benchmark is:

#TODO

It may vary by timespan considered and data vendors.

A variation of the VWAP benchmark is given by the \emph{Participation Weighted Price (PWP) benchmark}.
Where the weighting is with respect to the \emph{PWP shares}:
  
#TODO

Lastly the \emph{Relative Performance Measure}, which differs from the PnL metrics above. 
It is expressed as:

#TODO
}
\examples{

}
\references{
Kissell, R. \emph{The Science of Algorithmic Trading and Portfolio Management} (ISBN 978-0-12-401689-7)
}
\seealso{
\code{\link{initPortf}}, \code{\link{addTxn}}
}
\author{
Vito Lestingi
}
