% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tradeStats.R
\name{dailyTxnPL}
\alias{dailyTxnPL}
\alias{dailyEqPL}
\alias{dailyEqPL}
\title{generate daily Transaction Realized or Equity Curve P&L by instrument}
\usage{
dailyTxnPL(Portfolios, Symbols, drop.time = TRUE, incl.total = FALSE,
  envir = .blotter)

dailyEqPL(Portfolios, Symbols, drop.time = TRUE, incl.total = FALSE,
  envir = envir)
}
\arguments{
\item{Portfolios}{portfolio string}

\item{Symbols}{character vector of symbol strings}

\item{drop.time}{remove time component of POSIX datestamp (if any), default TRUE}

\item{incl.total}{if TRUE, add a column with the daily portfolio total P&L, default FALSE}
}
\value{
a multi-column \code{xts} time series, one column per symbol, one row per day
}
\description{
designed to collate information for high frequency portfolios
}
\details{
If you do not pass \code{Symbols}, then all symbols in the provided 
\code{Portfolios} will be used.

The daily P&L is calculated from \code{Net.Txn.Realized.PL} if by 
\code{dailyTxnPL} 
and from \code{Net.Trading.PL} by \code{dailyEqPL}
}
\seealso{
tradeStats
}
\author{
Brian G. Peterson
}
