% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEndEq.R
\name{getEndEq}
\alias{getEndEq}
\title{Retrieves the most recent value of the capital account}
\usage{
getEndEq(Account, Date)
}
\arguments{
\item{Account}{string identifier of account}

\item{Date}{last date to retrieve calculated equity for, as string}
}
\value{
Numeric value of the equity account
}
\description{
Retrieves the most recent value of the capital account
}

