% Generated by roxygen2 (4.0.2): do not edit by hand
\name{.getPortfAcct}
\alias{.getPortfAcct}
\title{get a protfolio in an account}
\usage{
.getPortfAcct(Account, Portfolio, Dates = NULL)
}
\arguments{
\item{Account}{account string}

\item{Portfolio}{portfolio string}

\item{Dates}{date subset as an xts style ISO 8601 string}
}
\description{
get a protfolio in an account
}

