\name{getPortfolio}
\alias{getPortfolio}
\title{get a portfolio object}
\usage{
  getPortfolio(Portfolio, Dates = NULL, envir = .blotter)
}
\arguments{
  \item{Portfolio}{string identifying portfolio}

  \item{Dates}{dates subset, not yet supported}

  \item{envir}{the environment to retrieve the portfolio
  from, defaults to .blotter}
}
\description{
  get a portfolio object
}

