\name{.getPosAvgCost}
\alias{.getPosAvgCost}
\title{Retrieves the most recent average cost of the position...}
\usage{.getPosAvgCost(Portfolio, Symbol, Date)}
\description{Retrieves the most recent average cost of the position}
\value{Numeric value of the average cost of the current position}
\arguments{\item{Portfolio}{a portfolio object containing transactions}
\item{Symbol}{an instrument identifier for a symbol included in the portfolio}
\item{Date}{timestamp as of which to have the most recent position}}
