% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getPosQty}
\alias{getPosQty}
\title{gets position at Date}
\usage{
getPosQty(Portfolio, Symbol, Date)
}
\arguments{
\item{Portfolio}{a string identifying a portfolio object containing transactions}

\item{Symbol}{an instrument identifier for a symbol included in the portfolio}

\item{Date}{timestamp as of which to have the most recent position}
}
\value{
Numeric value of the most recent position.
}
\description{
gets position at Date
}

