\name{getTxns}
\alias{getTxns}
\title{Gets the transactions and returns...}
\usage{getTxns(Portfolio, Symbol, Date)}
\description{Gets the transactions and returns}
\value{xts of transactions made in the Symbol during the time period given}
\arguments{\item{Portfolio}{a string identifying a portfolio object containing transactions}
\item{Symbol}{an instrument identifier for a symbol included in the portfolio}
\item{Date}{timestamp as of which to get transactions}}
