% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcsim.R
\name{hist.mcsim}
\alias{hist.mcsim}
\title{hist method for objects of type \code{mcsim}}
\usage{
\method{hist}{mcsim}(x, ..., normalize = TRUE)
}
\arguments{
\item{x}{object of type mcsim to plot}

\item{normalize}{TRUE/FALSE whether to normalize the hist by div, default TRUE}

\item{\dots}{any other passthrough parameters}
}
\description{
hist method for objects of type \code{mcsim}
}
\author{
Jasen Mackie, Brian G. Peterson
}

