\name{IBM}
\docType{data}
\alias{IBM}
\title{IBM stock price data for one month}
\description{
  This data set provides Open, High, Low, Close, Volume and
  AdjustedClose for IBM during the month of January 2007.
  It is included in the package to provide data for examples and tests. 
}
\usage{IBM}
\format{A \pkg{xts} object with six columns and a \code{Date}-class index.}
\source{The data set was retrieved from Yahoo! Finance.}
\keyword{datasets}
