% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndexAdjustedCost.R
\name{IndexAdjustedCost}
\alias{IndexAdjustedCost}
\title{Index Adjusted Cost}
\usage{
IndexAdjustedCost(ArrivalCost, IndexCost, beta)
}
\arguments{
\item{ArrivalCost}{value representing arrival cost}

\item{IndexCost}{value representing index cost}

\item{beta}{value represent underlying stock beta to index}
}
\description{
Index Adjusted Cost
}
\examples{
ac <- ArrivalCost(side = "Buy", p_avg = 100.50, p_0 = 100)
ic <- IndexCost(IndexVWAP = 50100, IndexArrivalCost = 50000)
iac <- IndexAdjustedCost(ac, ic, 0.7)

}
