% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndexAdjustedCost.R
\name{IndexCost}
\alias{IndexCost}
\title{Index Cost}
\usage{
IndexCost(IndexVWAP, IndexArrivalCost)
}
\arguments{
\item{IndexVWAP}{value representing the benchmark VWAP}

\item{IndexArrivalCost}{value representing the benchmark arrival price}
}
\description{
Index Cost
}
\examples{
IndexCost(IndexVWAP = 50100, IndexArrivalCost = 50000)

}
