% Generated by roxygen2 (4.0.2): do not edit by hand
\name{.initTxn}
\alias{.initTxn}
\title{Constructs the data container used to store transactions and resulting positions.}
\usage{
.initTxn(initDate = "1950-01-01", initPosQty = 0, ...)
}
\arguments{
\item{initDate}{date prior to the first close price given, used to contain initial account equity and initial position}

\item{initPosQty}{initial position, default is zero}

\item{\dots}{any other passthrough parameters}
}
\value{
Constructs multi-column xts object used to store transactions
}
\description{
The data series stored here will be an irregular time series.
}

