% Generated by roxygen2 (4.0.2): do not edit by hand
\name{is.portfolio}
\alias{is.portfolio}
\title{generic is.function for portfolio, will take either a string or an object}
\usage{
is.portfolio(x, ...)
}
\arguments{
\item{x}{an object or string to be tested as a portfolio}

\item{\dots}{any other passthru parameters}
}
\description{
If \code{x} is a string, \code{\link{getPortfolio}} will be called with
string \code{x} and tested.  Otherwise, the object passed will be tested.
}
\seealso{
\code{\link{getPortfolio}}
}

