% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iStar.R
\name{iStarPostTrade}
\alias{iStarPostTrade}
\title{The Kissell-Malamut \emph{I-Star} Market Impact model}
\usage{
iStarPostTrade(MktData, sessions, yrBizdays = 250, horizon = 30,
  xtsfy = FALSE, grouping = FALSE, targetIntervals, minDataPoints,
  paramsBounds, TxnData, side, ...)
}
\arguments{
\item{MktData}{A list of \code{xts} objects each representing a security market data. See 'Details'}

\item{sessions}{A character or a vector of character representing ISO time subsets to split each trading day in "sessions". If missing sessions are on a daily basis}

\item{yrBizdays}{A numeric value, the number of business days in a given year data refers to. Default is 250 days}

\item{horizon}{A numeric value, the number of sessions to compute the rolling variables over. Default is 30 days. See 'Details'}

\item{xtsfy}{A boolean specifying whether the rolling variables computed should become \code{xts} object with consistent dates}

\item{grouping}{A boolean or vector of booleans to specifying whether to group datapoints. Eventually, the second element specifies whether to average group values. Attention: the grouping may discard data. See 'Details'}

\item{targetIntervals}{A vector with named elements being 'ImSize', 'POV', 'Vol'. They have to be increasing sequences expressing the respective variable bounds, which are used to build datapoints groups. See 'Details'}

\item{minDataPoints}{A numeric value, the minimum number of datapoints a group should have to be included in the estimation process. Default is 25. See 'Details'}

\item{paramsBounds}{A matrix to provide model parameters bounds to pass to \code{nls}. Parameters are considered by row and columns represents lower and upper bounds, respectively}

\item{TxnData}{An \code{xts} object, with 'TxnPrice' and 'TxnQty' required columns. See 'Details'}

\item{side}{A numeric. Either 1 meaning 'buy' or -1 meaning 'sell'}

\item{...}{Passthrough parameters}
}
\value{
TODO: WIP 
\describe{
     \item{\code{}: }{}
     \item{\code{}: }{}
     \item{\code{}: }{}
}
}
\description{
The model is a cost allocation method to quantify the market impact of financial
transactions, depending on an agent order size relative to the market volume;  
in its authors words is theoretically based on the supply-demand principle, 
although it may be rather difficult to express ourselves precisely in these 
terms and even so our interpretations may differ by the several possible 
scenarios that take place into the market in response to imbalances.
}
\details{
The \code{MktData} input dataset must be a list, with items being market data
by security considered. It is suggested to name them to match the security they
refer to.
Each item is required to be an \code{xts} object which must have at least two 
columns, 'MktPrice' and 'MktQty'. For theoretical accuracy of the arrival price
it is recommended to input 'Bid' and 'Ask' columns as well. Similarly, providing 
a 'Reason' column allows to have trades classified by your preferred criterion; 
if it is not available the Lee-Ready tick test will be used to infere the trade 
direction.
If the \code{MktData} list provided has items with different number of observations,
then data considered will be only until to match the item with the smallest number
of observations. Also, beware that there is no strict timestamps matching (to avoid
unwanted restrictions on potentially mismatching intraday timestamps), therefore:
be careful to provide a data set which as at least the same number or unique 
days across all the securities included; be careful in checking that observations 
are on constistent dates across the full data set.
Our best suggestion is to use a data set within the same period and including 
the same number of days for each security involved in the analysis.

The \code{horizon} should be chosen according to the number of \code{sessions}
a trading day is splitted into.

Parameters \code{targetInterval} and \code{minDataPoints} regulate the grouping process.
\code{minDataPoints} of each group to let its datapoints belong to the estimation 
process dafaults to 25 datapoints, as suggested by the author. However, this parameter
largerly depends on the given dataset and on others parameters such as the \code{session} 
specifications.
TODO: add \code{targetInterval} default sequences, also specify bounds inclusion
}
\examples{

}
\references{
\emph{The Science of Algorithmic Trading and Portfolio Management} (Kissell, 2013), Elsevier Science.
\emph{A Practical Framework for Estimating Transaction Costs and Developing Optimal Trading Strategies to Achieve Best Execution} (Kissell, Glantz and Malamut, 2004), Finance Research Letters.
\emph{Inferring Trade Direction from Intraday Data} (Lee and Ready, 1991), The Journal of Finance.
}
\seealso{
\code{\link[PerformanceAnalytics]{Return.calculate}},
   \code{\link[PerformanceAnalytics]{sd.annualized}},
   \code{\link[stats]{nls}}
}
\author{
Vito Lestingi
}
