% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iStar.R
\name{iStarSensitivity}
\alias{iStarSensitivity}
\title{I-Star model sensitivity analysis}
\usage{
iStarSensitivity(object, paramsBounds, paramSteps, ...)
}
\arguments{
\item{object}{An object of class 'iStarEst' from the \code{iStarPostTrade} function # TODO: the class is not defined at present, nor checks are in place here}

\item{paramsBounds}{A matrix providing model parameters bounds to pass to \code{nls}. The same used in \code{iStarPostTrade} and defaults to the same values dicussed there}

\item{paramSteps}{A vector of named elements representing each parameter step size to build the parameters sequences with. See 'Details'}

\item{...}{Any other passthrough parameter}
}
\value{
A \code{list} with elements:
\describe{
     \item{\code{Params.Seqs}: }{A \code{list} of parameters sequences evaluated}
     \item{\code{nls.impact.fits}: }{A \code{list} of each model fitted with \code{nls}}
     \item{\code{Params.Sensitivity}: }{A \code{matrix} contining the results of the sensitivity analysis}
}
}
\description{
An helper function to provide I-Star model parameters sensitivity analysis.
}
\details{
The sensitivity analysis provided is a local one, with I-Star model paramaters  
being fixed one at a time. Then, for each fixed sequence of values provided for
a parameter, the nonlinear problem is solved to estimate the remaing four 
paramaters by means of nonlinear regression.

Results of the analysis are reported along with the corresponding \emph{residul
standard error} (RSE) of each model being fitted. This quantity is expressed 
in the same dependent variable unit and best fit paramaters should be such that 
this quantity is minimized.


Of course, \code{paramSteps} is related to \code{paramsBounds}. In particular, 
it should be stress that, provided step sizes will be used to built parameters 
sequences from the lower bound specified in \code{paramsBounds} until the last 
multiple of the upper bound provided in \code{paramsBounds} is reached, for each 
paramater and its respective bound values. In other words, \code{paramSteps}
is not allowed to have a sequence value that goes beyond the \code{paramsBounds}
specified upper bound.

\code{paramSteps} default is 50 for \eqn{a_1}, 0.1 for \eqn{a_2} and \eqn{a_3},
0.05 for \eqn{a_4} and 0.01 for \eqn{b_1}.
}
\note{
If paramaters fixed sequences values lead to nonlinear least squares estimation
failures, an \code{NA} is put in place of the other paramaters being estimated
and of the residual sum of squares, as they cannot be provided.
}
\references{
\emph{The Science of Algorithmic Trading and Portfolio Management} (Kissell, 2013), Elsevier Science.
}
\seealso{
\code{iStarPostTrade},
  \code{\link[stats]{nls}}
}
\author{
Vito Lestingi
}
