% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcsim.R
\name{mcsim}
\alias{mcsim}
\title{Monte Carlo simulate strategy results}
\usage{
mcsim(Portfolio, Account, n = 1000, ..., l = 20, use = c("equity",
  "txns"), pstart = 1)
}
\arguments{
\item{Portfolio}{string identifier of portfolio name}

\item{Account}{string identifier of account name}

\item{n}{number of simulations, default = 1000}

\item{l}{block length, default = 20}

\item{use}{determines whether to use 'daily' or 'txn' PL, default = "equity" ie. daily}

\item{pstart}{numeric identifier for which period in the backtest is day 1, useful for longet term MA strategies}

\item{\dots}{any other passthrough parameters}
}
\value{
a plot.xts object of simulated return streams

a hist object of sampled max drawdowns
}
\description{
Return bands of returns based on Monte Carlo simulations of back-test results
}
\note{
Requires boot package
}
\author{
Jasen Mackie, Brian G. Peterson
}
\seealso{
\code{\link{boot}}
}

