% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbo.R
\name{pbo.cscv}
\alias{pbo.cscv}
\title{Internal implementation of the Combinatorially Symmetric Cross Validation (CSCV) technique used by Marcos Lopez de Prado.}
\usage{
pbo.cscv(m = ret, s = 8, f = NA, threshold = 0, inf_sub = 6,
  allow_parallel = FALSE)
}
\arguments{
\item{m}{a \eqn{TxN} data frame of returns, where \eqn{T} is the samples per study and \eqn{N} is the number of studies.}

\item{s}{the number of subsets of \code{m} for CSCV combinations; 
must evenly divide \code{m}}

\item{f}{the function to evaluate a study's performance; required}

\item{threshold}{the performance metric threshold 
(e.g. 0 for Sharpe, 1 for Omega)}

\item{inf_sub}{infinity substitution value for reasonable plotting}

\item{allow_parallel}{whether to enable parallel processing, default FALSE}
}
\value{
object of class \code{pbo} containing list of PBO calculation results 
and settings
}
\description{
Internal function used by \code{blotter:::pbo} for computing the Probability of Backtest Overfitting.
}
\details{
This function performs the probability of backtest overfitting calculation
using a combinatorially-symmetric cross validation (CSCV) approach. We have ported most of the source code
from Matt Barry's implementation with his gracious consent. There will likely be a few extensions from Matt's
version, with the first being to make the function compatible with the \code{portfolios}
object in \code{blotter}.
}
\references{
Baily et al., "The Probability of Backtest Overfitting,"
\url{http://papers.ssrn.com/sol3/papers.cfm?abstract_id=2326253}

Baily et al., "Pseudo-Mathematics and Financial Charlatanism: The Effects of Backtest Overfitting on Out-of-Sample Performance,"
\url{https://papers.ssrn.com/sol3/papers.cfm?abstract_id=2308659}

Matt Barry's Vignette for the original \code{pbo} package
\url{https://cran.r-project.org/web/packages/pbo/vignettes/pbo.html}
}
\seealso{
\code{\link{pbo}}
}
\author{
Matt Barry, Jasen Mackie, Brian G. Peterson
}
\keyword{CSCV}
\keyword{PBO}
\keyword{backtest}
\keyword{overfitting}
\keyword{probability}
