\name{pennyPerShare}
\alias{pennyPerShare}
\title{example TxnFee cost function...}
\usage{pennyPerShare(TxnQty, TxnPrice)}
\description{example TxnFee cost function}
\arguments{\item{TxnQty}{total units (such as shares or contracts) transacted.  Positive values indicate a 'buy'; negative values indicate a 'sell'}
\item{TxnPrice}{price at which the transaction was done}}
