% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iStar.R
\name{plot.iStarEst}
\alias{plot.iStarEst}
\title{Plot method for object of type \code{iStarEst}}
\usage{
\method{plot}{iStarEst}(object, xVar, fixVars, fixVals, OrdData = NULL,
  ...)
}
\arguments{
\item{object}{An object of class \code{iStarEst} from \code{iStarPostTrade}}

\item{xVar}{A character specifying the variable to plot the market impact against, one of 'Size' (default), 'POV' or 'AnnualVol'}

\item{fixVars}{A vector of character specifying the variable to fix. A couple of 'Size', 'POV' (default) or 'AnnualVol' (default)}

\item{fixVals}{A vector with named elements representing the values to fix \code{fixVars} at, in decimal units}

\item{OrdData}{Same as in \code{iStarPostTrade}}

\item{...}{Any other passthrough parameter}
}
\value{
In the produced plot, the chosen \code{xVar} and \code{fixVars} are expressed 
in percentages terms, as they are most commonly used in reports.
}
\description{
The main scope of present function is to reproduce what Kissell refers to as 
\emph{cost curves}. In these curves estimated market impact is plotted against 
the order size for a given value of the volatility or against the volatility 
for a given order size. Likewise, we can plot the estimated market impact against 
sequences of 'POV' or 'AnnualVol' with the other two respective variables being 
fixed.
}
\details{
The plots are mainly provided for exaplanatory purposes. They are a qualitatite
mean to extend an order scrutiny to the market impact estimated cost shape under 
different values of the relevant variables. Cost curves are expected to possess 
a concave shape.

It shall be stressed that even when actual custom order data will be provided,
we do not use it to compute variables underlying the cost curve. Instead, the 
selected variable is varied along a given sequence and for provided values of
the other two variables.

 
For comparisons consistency, the \code{OrdData} must be the one passed to 
\code{iStarPostTrade} to compute I-Star model impact estimates.
}
\examples{

}
\references{
\emph{The Science of Algorithmic Trading and Portfolio Management} (Kissell, 2013), Elsevier Science.
}
\seealso{
\code{\link{iStarPostTrade}}
}
\author{
Vito Lestingi
}
