% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iStar.R
\name{plot.iStarSens}
\alias{plot.iStarSens}
\title{Plot method for object of type \code{iStarSens}}
\usage{
\method{plot}{iStarSens}(x, fix, ...)
}
\arguments{
\item{x}{An object of class \code{iStarSens} from \code{iStarSensitivity}}

\item{fix}{A character, the name of the parameter to fix. One of 'a_1', 'a_2', 'a_3', 'a_4' or 'b_1'}

\item{...}{Any other passthrough parameter}
}
\description{
S3 method to conveniently plot \code{iStarSensitivity} results on I-Star model 
parameters sensitivity.
}
\details{
The chosen parameter, or better its given sequence specified in the \code{iStarSensitivity}
call, is plotted against the \code{RSE} of the models fitted on the remaining 
parameters.
}
\references{
\emph{The Science of Algorithmic Trading and Portfolio Management} (Kissell, 2013), Elsevier Science.
}
\author{
Vito Lestingi
}
