% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcsim.R
\name{plot.mcsim}
\alias{plot.mcsim}
\title{plot method for objects of type \code{mcsim}}
\usage{
\method{plot}{mcsim}(x, y, ..., normalize = TRUE)
}
\arguments{
\item{x}{object of type 'mcsim' to plot}

\item{y}{not used, to match generic signature, may hold overlay data in the future}

\item{\dots}{any other passthrough parameters}

\item{normalize}{TRUE/FALSE whether to normalize the plot by initEq, default TRUE}
}
\description{
plot method for objects of type \code{mcsim}
}
\seealso{
\code{\link{mcsim}}
}
\author{
Jasen Mackie, Brian G. Peterson
}
