% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txnsim.R
\name{plot.txnsim}
\alias{plot.txnsim}
\title{plot method for objects of type 'txnsim'}
\usage{
\method{plot}{txnsim}(x, y, ...)
}
\arguments{
\item{x}{object of type 'txnsim' to plot}

\item{y}{not used, to match generic signature, may hold overlay data in the future}

\item{\dots}{any other passthrough parameters}
}
\description{
plot method for objects of type 'txnsim'
}
\author{
Jasen Mackie, Brian G. Peterson
}
\seealso{
\code{\link{txnsim}}
}

