% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchTradePerf.R
\name{plot.txnsPerf}
\alias{plot.txnsPerf}
\title{Plot method for object of type \code{txnsPerf}}
\usage{
\method{plot}{txnsPerf}(object, benchmark, legend.loc, ...)
}
\arguments{
\item{object}{Object of type \code{txnsPerf} to plot}

\item{benchmark}{String identifying the benchmark used to produce the \code{txnsPerf} object}

\item{legend.loc}{String specifying the position of second panel legend}

\item{...}{Any other passthrough parameters}
}
\description{
Plot method for object of type \code{txnsPerf}
}
\examples{

plot(benchTradeBench, benchmark = 'TradeBench')
plot(benchMktBenchOpen, benchmark = 'MktBench')
plot(benchMktBenchClose, benchmark = 'MktBench')
plot(benchVWAPfull, benchmark = 'VWAP')
plot(benchPWP, benchmark = 'PWP')
plot(benchRPM, benchmark = 'RPM')

}
\seealso{
\code{\link[xts]{plot.xts}}
}
\author{
Vito Lestingi
}
