% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchTradeStats.R
\name{print.txnsStats}
\alias{print.txnsStats}
\title{Print method for object of class \code{txnsStats}}
\usage{
\method{print}{txnsStats}(x, ...)
}
\arguments{
\item{x}{Object of class \code{txnsStats} to print}

\item{...}{Any other passthrough parameters}
}
\value{
The \code{txnsStats} input object without its first element.
}
\description{
To prevent cluttering the console with \code{benchTradeStats()} output other
than statistical testing and reporting main purposes.
}
