% Generated by roxygen2 (4.0.2): do not edit by hand
\name{put.account}
\alias{put.account}
\title{put a account object in .blotter env}
\usage{
put.account(account.st, account, envir = .blotter)
}
\arguments{
\item{account.st}{string identifying account}

\item{account}{account object}

\item{envir}{the environment to save the account object in, defaults to .blotter}
}
\description{
put a account object in .blotter env
}

