% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txnsim.R
\name{quantile.txnsim}
\alias{quantile.txnsim}
\title{quantile method for objects of type \code{txnsim}}
\usage{
\method{quantile}{txnsim}(x, ...)
}
\arguments{
\item{x}{object of type 'txnsim' to produce replicate quantiles}

\item{\dots}{any other passthrough parameters to \code{\link{quantile}}}
}
\description{
calculates quantiles of cumulative P&L of the simulated strategies
}
\author{
Jasen Mackie, Brian G. Peterson
}
