% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcsim.R
\name{summary.mcsim}
\alias{summary.mcsim}
\alias{print.mcsim}
\title{summary and print methods for objects of type \code{mcsim}}
\usage{
\method{summary}{mcsim}(object, ..., normalize = TRUE)

\method{print}{mcsim}(x, ..., normalize = TRUE)
}
\arguments{
\item{object}{object of type 'mcsim' to produce a sample and backtest summary}

\item{\dots}{any other passthrough parameters}

\item{normalize}{TRUE/FALSE whether to use normalized percent-based summary stats, default TRUE}

\item{x}{an object of type 'mcsim'}
}
\description{
summary and print methods for objects of type \code{mcsim}
}
\author{
Jasen Mackie, Brian G. Peterson
}
