% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbo.R
\name{summary.pbo}
\alias{summary.pbo}
\title{summary method for objects of type \code{pbo}}
\usage{
\method{summary}{pbo}(object, ...)
}
\arguments{
\item{object}{object of type 'pbo' to summarise}

\item{...}{\dots any other passthrough parameters}
}
\description{
summary method for objects of type \code{pbo}
}
