% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchTradePerf.R
\name{summary.txnsPerf}
\alias{summary.txnsPerf}
\title{Summary method for object of type \code{txnsPerf}}
\usage{
\method{summary}{txnsPerf}(object, ...)
}
\arguments{
\item{object}{Object of type \code{txnsPerf} to plot}

\item{...}{Any other passthrough parameters}
}
\description{
Summary method for object of type \code{txnsPerf}
}
