% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trdPrcPerf.R
\name{trdPrcPerf}
\alias{trdPrcPerf}
\title{Trading Price Performance}
\usage{
trdPrcPerf(Portfolio, Symbol, side = "Buy")
}
\arguments{
\item{Portfolio}{A portfolio name that points to a portfolio object structured with initPortf()}

\item{Symbol}{An instrument identifier for a symbol included in the portfolio, e.g., "IBM"}

\item{side}{string identifying either "Buy" or "Sell"}
}
\value{
A numeric value in basis points
}
\description{
Trading Price Performance or trading PnL is the measure of cost during trading and tell whether the investor 
did better or worse as compared to arrival price.
}
\examples{

}
