% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txnsim.R
\name{txnsim.portfs}
\alias{txnsim.portfs}
\title{convenience function to generate portfolios for txnsim replicates}
\usage{
txnsim.portfs(Portfolio, replacement, n, symbols, initDate, currency, initEq)
}
\arguments{
\item{Portfolio}{string identifying a portfolio}

\item{replacement}{sample with or without replacement, default TRUE}

\item{n}{number of simulations, default = 100}

\item{symbols}{character vector of symbol names}

\item{initDate}{initialization Date to use for replicate portfolios}

\item{currency}{base currency for replicate portfolios}

\item{initEq}{initial equity to use for replicate portfolios}
}
\description{
If you have a txnsim object and market data, you should be able to rebuild 
the replicate portfolios. This function creates all those portfolios.
}
\seealso{
\code{\link{txnsim}}, \code{\link{txnsim.txns}}
}
