% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txnsim.R
\name{txnsim}
\alias{txnsim}
\title{Monte Carlo analysis of transactions}
\usage{
txnsim(Portfolio, n = 10, replacement = TRUE, tradeDef = "flat.to.flat",
  ...)
}
\arguments{
\item{Portfolio}{string identifying a portfolio}

\item{n}{number of simulations, default = 100}

\item{replacement}{sample with or without replacement, default TRUE}

\item{tradeDef}{string to determine which definition of 'trade' to use. See \code{\link{tradeStats}}}

\item{\dots}{any other passthrough parameters}
}
\value{
a list object of class 'txnsim' containing:
\itemize{
  \item{\code{backtest.trades}:}{list by symbol containing trade start, quantity, duration from the original backtest}
  \item{\code{replicates}:}{a list by symbol containing all the resampled start,quantity, duration time series replicates}
  \item{\code{transactions}:}{a list by symbol for each replicate of the Txn object passed to \code{\link{addTxns}}}
  \item{\code{initeq}:}{a numeric variable containing the initEq of the portfolio, for starting portfolio value}
  \item{\code{seed}:}{ the value of \code{.Random.seed} for replication, if required}
  \item{\code{call}:}{an object of type \code{call} that contains the \code{txnsim} call}
}

Note that this object and its slots may change in the future.
Slots \code{replicates},\code{transactions}, and \code{call} are likely
to exist in all future versions of this function, but other slots may be added
and removed as \code{S3method}'s are developed.
}
\description{
Monte Carlo analysis of transactions
}
\examples{
\dontrun{

  n <- 10

  ex.txnsim <- function(Portfolio,n,replacement) {
    out <- txnsim(Portfolio,n,replacement)
    for (i in 1:n){
      p<-paste('txnsim',Portfolio,i,sep='.')
      updatePortf(p)
      symbols<-names(getPortfolio(p)$symbols)
      for(symbol in symbols) {
        dev.new()
        chart.Posn(p,symbol)
      }
    }
    out
  }

  demo('longtrend',ask=FALSE)
  lt.nr <- ex.txnsim('longtrend',n, replacement = FALSE)
  lt.wr <- ex.txnsim('longtrend',n, replacement = TRUE)

  require('quantstrat')
  demo('rsi',ask=FALSE)
  rsi.nr <- ex.txnsim('RSI',n, replacement = FALSE)
  rsi.wr <- ex.txnsim('RSI',n, replacement = TRUE)

} #end dontrun

}
\author{
Jasen Mackie, Brian G. Peterson
}

