% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txnsim.R
\name{txnsim.txns}
\alias{txnsim.txns}
\title{convenience function to create transactions from txnsim replicates}
\usage{
txnsim.txns(reps, Portfolio, replacement, n, ...)
}
\arguments{
\item{reps}{replicates slot from txnsim object}

\item{Portfolio}{string identifying a portfolio}

\item{replacement}{sample with or without replacement, default TRUE}

\item{n}{number of simulations, default = 100}

\item{\dots}{any other passthrough parameters, most usefully \code{env} and \code{prefer}}
}
\value{
a list by symbol of txns suitable for calling \code{\link{addTxns}}
}
\description{
If you have a txnsim object and market data, you should be able to rebuild
the replicate portfolios.
}
\seealso{
\code{\link{txnsim}}, \code{\link{txnsim.txns}} , \code{\link{addTxns}}
}
