\name{addPortfInstr}
\alias{addPortfInstr}
\title{add an instrument to a portfolio...}
\usage{addPortfInstr(Portfolio, symbols, ...)}
\description{add an instrument to a portfolio}
\details{thanks to WolfGang Wu for making this function more usable}
\arguments{\item{Portfolio}{portfolio identifier string}
\item{symbols}{character vector of symbols to add to the portfolio}
\item{\dots}{any other passthru parameters}}
