% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndexAdjustedCost.R
\name{ArrivalCost}
\alias{ArrivalCost}
\title{Arrival Cost}
\usage{
ArrivalCost(side = "Buy", p_avg = NULL, p_0 = NULL)
}
\arguments{
\item{side}{string identifying either "Buy" or "Sell"}

\item{p_avg}{value representing average executoin price}

\item{p_0}{value representing arrival price}
}
\description{
The Arrival Cost (or Trading Cost) is the difference between the
average execution price and the price of the asset at the time the
order was entered into the market, ie. the Arrival Price.
}
\details{
\deqn{ArrivalCost_bp} = Side \dot \frac{P_avg - P_0 }{P_0} \dot {10^4}_bp
\deqn{{p^s} = Pr( |r| > t-ratio)}{p^s = Pr(|r|>t-ratio)}
}
\examples{
ArrivalCost(side = "Buy", p_avg = 100.50, p_0 = 100)

}
