% Generated by roxygen2 (4.0.2): do not edit by hand
\name{calcPortfWgt}
\alias{calcPortfWgt}
\title{Calculates the portfolio weights for positions within a given portfolio.}
\usage{
calcPortfWgt(Portfolio, Symbols = NULL, Dates = NULL,
  denominator = c("Gross.Value", "Net.Value", "Long.Value", "Short.Value"),
  Account)
}
\arguments{
\item{Portfolio}{a portfolio object structured with initPortf()}

\item{Symbols}{an instrument identifier for a symbol included in the portfolio}

\item{Dates}{dates to return the calculation over formatted as xts range}

\item{denominator}{string describing the deniminator, see Description}

\item{Account}{an Account object containing Portfolio summaries}
}
\value{
xts timeseries object with weights by date in rows and symbolname in columns
}
\description{
Portfolio weights may be calculated differently depending on their use.
By default, this function uses denominator of 'Gross.Value', the second most common
option will likely be 'Net.Value'.
For separating long and short weights, 'Long.Value' and 'Short.Value' may be
needed as denominators.
}

