% Generated by roxygen2 (4.0.2): do not edit by hand
\name{chart.ME}
\alias{chart.ME}
\title{Chart Maximum Adverse/Favorable Excursion}
\usage{
chart.ME(Portfolio, Symbol, type = c("MAE", "MFE"), scale = c("cash",
  "percent", "tick"), ...)
}
\arguments{
\item{Portfolio}{string identifying the portfolio to chart}

\item{Symbol}{string identifying the symbol to chart. If missing, the first symbol found in the \code{Portfolio} portfolio will be used}

\item{type}{string specifying MAE (Adverse) or MFE (Favourable) chart type}

\item{scale}{string specifying 'cash', or 'percent' for percentage of investment, or 'tick'}

\item{\dots}{any other passthrough parameters, in particular includeOpenTrades (see perTradeStats())}
}
\description{
Produces a scatterplot with one point per trade, with x-axis: absolute
value of Drawdown (Adverse), or Run Up (Favourable),
and y-axis: absolute value of Net Profit or Loss
}
\author{
Jan Humme
}
\references{
Tomasini, E. and Jaekle, U. \emph{Trading Systems - A new approach to system development and portfolio optimisation} (ISBN 978-1-905641-79-6), section 3.5
}
\seealso{
\code{\link{perTradeStats}} for the calculations used by this chart,
and \code{\link{tradeStats}} for a summary view of the performance
}

