\name{chart.Posn}
\alias{chart.Posn}
\title{chart trades against market data, position through time, and equity curve...}
\usage{chart.Posn(Portfolio, Symbol, Dates, ...)}
\description{chart trades against market data, position through time, and equity curve}
\arguments{\item{Portfolio}{string identifying the portfolio to chart}
\item{Symbol}{string identifying the symbol to chart}
\item{Dates}{date range, currently not used}
\item{...}{any other passthru parameters (typically parameters to \code{chart_Series})}}
