% Generated by roxygen2 (4.0.2): do not edit by hand
\name{chart.Posn}
\alias{chart.Posn}
\title{Chart trades against market data, position through time, and cumulative P\&L}
\usage{
chart.Posn(Portfolio, Symbol, Dates = NULL, ..., TA = NULL)
}
\arguments{
\item{Portfolio}{string identifying the portfolio to chart}

\item{Symbol}{string identifying the symbol to chart. If missing, the first symbol found in the \code{Portfolio} portfolio will be used}

\item{Dates}{xts ISO 8601 style subsetting}

\item{TA}{a string defining a technical indicator function that will be applied to the chart, using \code{\link{eval}}}

\item{\dots}{any other passthru parameters to \code{\link[quantmod]{chart_Series}}}
}
\description{
Produces a three-panel chart of time series charts that contains prices and transactions in the top panel, the resulting position in the second, and a cumulative profit-loss line chart in the third.
}
\note{
Expect changes to this function, since the underlying charts are experimental functions in quantmod.
}

