\name{getAccount}
\alias{getAccount}
\title{Get an account object from the...}
\usage{getAccount(Account)}
\description{Get an account object from the .blotter environment}
\details{Retrieves an account object from the \code{.blotter} environment.  Useful for local examination or charting, or storing interim results for later reference.}
\value{Account object}
\arguments{\item{Account}{string identifier for the account}}
