\name{getAccount}
\alias{getAccount}
\title{get an account object from the environment for examination or manipulation...}
\usage{getAccount(Account, Dates)}
\description{get an account object from the environment for examination or manipulation}
\value{Account object}
\arguments{\item{Account}{string identifier for the account}
\item{Dates}{}}
