\name{getAccount}
\alias{getAccount}
\title{Get an account object from the .blotter environment}
\usage{
  getAccount(Account)
}
\arguments{
  \item{Account}{string identifier for the account}
}
\value{
  Account object
}
\description{
  Retrieves an account object from the \code{.blotter}
  environment.  Useful for local examination or charting,
  or storing interim results for later reference.
}

