% Generated by roxygen2 (4.0.2): do not edit by hand
\name{.getBySymbol}
\alias{.getBySymbol}
\title{Retrieves calculated attributes for each position in the portfolio}
\usage{
.getBySymbol(Portfolio, Attribute, Dates = NULL, Symbols = NULL,
  native = FALSE)
}
\arguments{
\item{Portfolio}{a portfolio object containing transactions}

\item{Attribute}{column name to be assembled for each symbol}

\item{Dates}{dates to return the calculation over formatted as xts range}

\item{Symbols}{an instrument identifier for a symbol included in the portfolio}

\item{native}{whether to use the currency of the portfolio, or convert, default FALSE}
}
\value{
regular xts object of values by symbol
}
\description{
Retrieves calculated attributes for each position in the portfolio
from the posPL table.  Assembles into a symbol-by-time table useful
for graphing or calculations
}
\details{
items typically include things like
'Pos.Qty', 'Pos.Value', 'Txn.Value', 'Realized.PL', 'Unrealized.PL',or 'Trading.PL'
}

