\name{getEndEq}
\alias{getEndEq}
\title{Retrieves the most recent value of the capital account...}
\usage{getEndEq(Account, Date)}
\description{Retrieves the most recent value of the capital account}
\value{Numeric value of the equity account}
\arguments{\item{Account}{string identifier of account}
\item{Date}{last date to retrieve calculated equity for, as string}}
