% Generated by roxygen2 (4.0.2): do not edit by hand
\name{.getPosAvgCost}
\alias{.getPosAvgCost}
\title{Retrieves the most recent average cost of the position}
\usage{
.getPosAvgCost(Portfolio, Symbol, Date)
}
\arguments{
\item{Portfolio}{a portfolio object containing transactions}

\item{Symbol}{an instrument identifier for a symbol included in the portfolio}

\item{Date}{timestamp as of which to have the most recent position}
}
\value{
Numeric value of the average cost of the current position
}
\description{
Retrieves the most recent average cost of the position
}

