% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndexAdjustedCost.R
\name{IndexAdjustedCost}
\alias{IndexAdjustedCost}
\title{Index Adjusted Cost}
\usage{
IndexAdjustedCost(ArrivalCost, IndexCost, beta)
}
\arguments{
\item{ArrivalCost}{value representing arrival cost}

\item{IndexCost}{value representing index cost}

\item{beta}{value representing stock K's sensitivity to the index}
}
\value{
The Index Adjusted Cost
}
\description{
A market or index adjusted metric accounting for price trend in a stock
due to the overall market.
}
\examples{

}
