% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndexAdjustedCost.R
\name{IndexCost}
\alias{IndexCost}
\title{Index Cost}
\usage{
IndexCost(Portfolio, IndexSymbol)
}
\arguments{
\item{Portfolio}{A portfolio name that points to a portfolio object structured with initPortf()}

\item{IndexSymbol}{An instrument identifier for an index used for adjusting a stock's Arrival Cost (typicall ETF)}
}
\value{
A numeric value in basis points
}
\description{
The Index Cost is the difference between the
average execution price and the price of the Index (or ETF) at the time the
order was entered into the market, ie. the Arrival Price.
}
\details{
To get the index VWAP of one does not have the underlying trades and weightings,
a commonly used proxy is an ETF tracking the relevant index.

Positive means a cost, negative means a saving.
}
\examples{

}
