% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcsim.R
\name{mcsim}
\alias{mcsim}
\title{Monte Carlo simulate strategy results}
\usage{
mcsim(
  Portfolio = NULL,
  Account = NULL,
  n = 1000,
  replacement = TRUE,
  ...,
  use = c("equity", "txns", "returns", "cash"),
  l = 1,
  varblock = FALSE,
  gap = 1,
  CI = 0.95,
  cashPL = NULL
)
}
\arguments{
\item{Portfolio}{string identifier of portfolio name}

\item{Account}{string identifier of account name}

\item{n}{number of simulations, default = 1000}

\item{replacement}{sample with or without replacement, default TRUE}

\item{\dots}{any other passthrough parameters}

\item{use}{determines whether to use 'equity', 'txn', 'returns', or 'cash',
Profit and Loss , default = "equity" ie. daily}

\item{l}{block length, default = 1}

\item{varblock}{boolean to determine whether to use variable block length, default FALSE}

\item{gap}{numeric number of periods from start of series to start on, to eliminate leading NA's}

\item{CI}{numeric specifying desired Confidence Interval used in hist.mcsim(), default 0.95}

\item{cashPL}{optional regular xts object of cash P&L if \code{use='cash'} and you don't want to use a blotter Portfolio to get P&L.}
}
\value{
a list object of class 'mcsim' containing:
\itemize{
  \item{\code{replicates}:}{an xts object containing all the resampled time series replicates}
  \item{\code{percreplicates}:}{an xts object containing all the resampled time series replicates in percent}
  \item{\code{dailypl}:}{an xts object containing daily P&L from the original backtest}
  \item{\code{percdailypl}:}{an xts object containing daily P&L in percent from the original backtest}
  \item{\code{initeq}:}{a numeric variable containing the initEq of the portfolio, for starting portfolio value}
  \item{\code{num}:}{a numeric variable reporting the number of replicaes in the simulation}
  \item{\code{length}:}{a numeric variable reporting the block length used in the simulation, if any}
  \item{\code{samplestats}:}{a numeric dataframe of various statistics for each replicate series}
  \item{\code{percsamplestats}:}{a numeric dataframe of various statistics for each replicate series in percent}
  \item{\code{original}:}{a numeric dataframe of the statistics for the original series}
  \item{\code{percoriginal}:}{a numeric dataframe of the statistics for the original series in percent terms}
  \item{\code{stderror}:}{a numeric dataframe of the standard error of the statistics for the replicates}
  \item{\code{percstderror}:}{a numeric dataframe of the standard error of the statistics for the replicates in percent}
  \item{\code{CI}:}{numeric specifying desired Confidence Interval used in hist.mcsim(), default 0.95}
  \item{\code{CIdf}:}{a numeric dataframe of the Confidence Intervals of the statistics for the bootstrapped replicates}
  \item{\code{CIdf_perc}:}{a numeric dataframe of the Confidence Intervals of the statistics for the bootstrapped replicates in percent}
  \item{\code{w}:}{a string containing information on whether the simulation is with or without replacement}
  \item{\code{use}:}{ a string with the value of the 'use' parameter, for checking later}
  \item{\code{seed}:}{ the value of \code{.Random.seed} for replication, if required}
  \item{\code{call}:}{an object of type \code{call} that contains the \code{mcsim} call}
}

If \code{use='cash'}, you must also supply a daily (or other regular frequency) 
cash P&L xts object in the \code{dailyPL} argument.

Note that this object and its slots may change in the future.
Slots \code{replicates},\code{dailypl},\code{initeq}, and \code{call} are likely
to exist in all future versions of this function, but other slots may be added
and removed as \code{S3method}'s are developed.
}
\description{
This function resamples the daily transaction, cash equity, or percent-return
P&L from a portfolio of trading results.  It may be applied to both real
(post-trade) and backtested transactions.
}
\details{
The general argument here is that you can compare the performance of real
portfolio equity or a backtest equity curve to a sampled version of the same.

We've chosen to use daily frequency for the resampling period. If your holding
period is longer than one day, on average, the samples will increase
variability in the overall path.  If your average holding period is shorter
than a day, the \code{\link{mcsim}} function will still provide a useful
benchmark for comparing to other strategies, but you may additionally wish to
sample round turn trades, as provided in (TODO: add link once function exists).

A few of the arguments and methods probably deserve more discussion as well.

\code{use} describes the method to use to generate the initial daily Profit and Loss  to
be sampled from.  The options are:
\itemize{
  \item{equity}{will use daily portfolio cash Profit and Loss}
  \item{txn}{will use transaction \code{Net.Trading.PL}}
}

Sampling may be performed either with or without replacement.
\itemize{
  \item{without replacement}{If sampled **without** replacement, the replicated
  equity curves will use all the same data as the input series, only reordered.
  This will lead to all the replicates having identical final Profit and Loss 
  and mean returns, but different paths along the way.}
  \item{with replacement}{If sampled **with** replacement, individual
  observations may be reused.  This will tend to create more variability than
  replicates without replacement.}
}

If the post-trade or backtested equity curve exhibits autocorrelation, runs,
streaks, etc. it may be advantageous to utilize a block resampling method.
The length of the block "\code{l}" may be fixed or variable.
If a \code{varblock} method is used, the distribution of block lengths will
be uniform random for \code{replacement=FALSE} and geometric random for
\code{replacement=TRUE}.  By sampling blocks, the resampled returns will
contain more of the structure of the original series.  If \code{varblock=TRUE},
the blocks will be of variable length, centered around \code{l}.
}
\note{
Requires boot package
}
\examples{
\dontrun{

demo('longtrend', ask=FALSE)

nrsim <- mcsim("longtrend", "longtrend", n=1000, replacement=FALSE, l=1, gap=10, CI=0.95)
nrblocksim.75 <- mcsim("longtrend", "longtrend", n=1000, replacement=FALSE, l=10, gap=10, CI=0.75)
rsim <- mcsim("longtrend", "longtrend", n=1000, replacement=TRUE, l=1, gap=10)
varsim <- mcsim("longtrend", "longtrend", n=1000, replacement=TRUE, l=10, varblock=TRUE, gap=10)
nrvarsim <- mcsim("longtrend", "longtrend", n=1000, replacement=FALSE, l=10, varblock=TRUE, gap=10)

quantile(varsim)
quantile(nrsim)
quantile(nrvarsim)

summary(varsim, normalize=FALSE)
summary(nrsim)
summary(nrvarsim)

plot(nrsim, normalize=TRUE)
plot(nrsim, normalize=FALSE)
plot(varsim)
plot(rsim)
hist(nrblocksim.75)
hist(rsim)
hist(varsim)

} #end dontrun

}
\seealso{
\code{\link{boot}}
  \code{\link{plot.mcsim}}
  \code{\link{hist.mcsim}}
}
\author{
Jasen Mackie, Brian G. Peterson
}
