% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbo.R
\name{pbo}
\alias{pbo}
\title{Probability of backtest overfitting}
\usage{
pbo(portfolios, ..., strategy = NULL, trials = NULL, audit = NULL,
  env = .GlobalEnv)
}
\arguments{
\item{portfolios}{string name of portfolio, or optionally a vector of portfolios, see DETAILS}

\item{...}{any other passthrough parameters}

\item{strategy}{optional strategy specification that would contain more information on the process, default NULL}

\item{trials}{optional number of trials,default NULL}

\item{audit}{optional audit environment containing the results of parameter optimization or walk forward, default NULL}

\item{env}{optional environment to find market data in, if required.}
}
\value{
an object of type \code{pbo.cscv} containing:
}
\description{
Performs the probability of backtest overfitting computations.
}
\details{
This function performs the probability of backtest overfitting calculation
using a combinatorially-symmetric cross validation (CSCV) approach.
}
\author{
Jasen Mackie, Brian G. Peterson
}
