\name{pennyPerShare}
\alias{pennyPerShare}
\title{example TxnFee cost function...}
\usage{pennyPerShare(TxnQty)}
\description{example TxnFee cost function}
\arguments{\item{TxnQty}{total units (such as shares or contracts) transacted.  Positive values indicate a 'buy'; negative values indicate a 'sell'
This is an example intended to demonstrate how a cost function could be used in place of a flat numeric fee.}}
