% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pennyPerShare}
\alias{pennyPerShare}
\title{Example TxnFee cost function}
\usage{
pennyPerShare(TxnQty, ...)
}
\arguments{
\item{TxnQty}{total units (such as shares or contracts) transacted.  Positive values indicate a 'buy'; negative values indicate a 'sell'
This is an example intended to demonstrate how a cost function could be used in place of a flat numeric fee.}

\item{\dots}{any other passthrough parameters}
}
\description{
Example TxnFee cost function
}

