% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchTradePerf.R
\name{plot.txnsPerf}
\alias{plot.txnsPerf}
\title{Plot method for object of type \code{txnsPerf}}
\usage{
\method{plot}{txnsPerf}(x, benchmark, legend.loc, ...)
}
\arguments{
\item{x}{Object of type \code{txnsPerf} to plot}

\item{benchmark}{String identifying the benchmark used to produce the \code{txnsPerf} object}

\item{legend.loc}{String specifying the position of second panel legend}

\item{...}{Any other passthrough parameters}
}
\description{
Plot method for object of type \code{txnsPerf}
}
\seealso{
\code{\link[xts]{plot.xts}}
}
\author{
Vito Lestingi, Jasen Mackie
}
