% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcsim.R
\name{quantile.mcsim}
\alias{quantile.mcsim}
\title{quantile method for objects of type \code{mcsim}}
\usage{
\method{quantile}{mcsim}(x, ..., normalize = TRUE)
}
\arguments{
\item{x}{object of type 'mcsim' to produce replicate quantiles}

\item{normalize}{TRUE/FALSE whether to normalize the plot by initEq, default TRUE}

\item{\dots}{any other passthrough parameters}
}
\description{
quantile method for objects of type \code{mcsim}
}
\author{
Jasen Mackie, Brian G. Peterson
}

