% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcsim.R
\name{summary.mcsim}
\alias{summary.mcsim}
\title{summary method for objects of type \code{mcsim}}
\usage{
summary.mcsim(object, ..., normalize = TRUE)
}
\arguments{
\item{object}{object of type 'mcsim' to produce a sample and backtest summary}

\item{normalize}{TRUE/FALSE whether to use normalized percent-based summary stats, default TRUE}

\item{\dots}{any other passthrough parameters}
}
\description{
summary method for objects of type \code{mcsim}
}
\author{
Jasen Mackie, Brian G. Peterson
}

