% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbo.R
\name{summary.pbo}
\alias{summary.pbo}
\title{summary method for objects of type \code{pbo}}
\usage{
summary.pbo(object, ...)
}
\arguments{
\item{object}{}

\item{...}{}
}
\description{
summary method for objects of type \code{pbo}
}
