% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txnsim.R
\name{summary.txnsim}
\alias{summary.txnsim}
\alias{print.txnsim}
\title{summary and print methods for objects of type txnsim}
\usage{
\method{summary}{txnsim}(object, ...)

\method{print}{txnsim}(x, ...)
}
\arguments{
\item{object}{an object of type txnsim}

\item{...}{any other passthrough parameters}

\item{x}{an object of type txnsim}
}
\description{
summary and print methods for objects of type txnsim
}
