\name{updateAcct}
\alias{updateAcct}
\title{Constructs the equity account calculations from the portfolio data and...}
\usage{updateAcct(name="default", Dates)}
\description{Constructs the equity account calculations from the portfolio data and
corresponding close prices.}
\arguments{\item{name}{string identifying account}
\item{Dates}{Dates from which to calculate equity account}}
