% Generated by roxygen2 (4.0.2): do not edit by hand
\name{updateAcct}
\alias{updateAcct}
\title{Constructs the equity account calculations from the portfolio data and
corresponding close prices.}
\usage{
updateAcct(name = "default", Dates = NULL)
}
\arguments{
\item{name}{string identifying account}

\item{Dates}{Dates from which to calculate equity account}
}
\description{
Constructs the equity account calculations from the portfolio data and
corresponding close prices.
}

