% Generated by roxygen2 (4.0.2): do not edit by hand
\name{updatePortf}
\alias{updatePortf}
\title{update Portfilio P&L over a Dates range}
\usage{
updatePortf(Portfolio, Symbols = NULL, Dates = NULL, Prices = NULL,
  Interval = Interval, ...)
}
\arguments{
\item{Portfolio}{string identifying a portfolio}

\item{Symbols}{character vector identifying symbols to update the portfolio for, default NULL}

\item{Dates}{optional xts-style ISO-8601 time range to run updatePortf over, default NULL (will use times from Prices)}

\item{Prices}{optional xts object containing prices and timestamps to mark the book on, default NULL}

\item{Interval}{optional character string, containing one of "millisecond" (or "ms"), "microsecond" (or "us"),
"second", "minute", "hour", "day", "week", "month", "quarter", or "year".  This can optionally be preceded by
a positive integer, or followed by "s".}

\item{\dots}{any other passthrough parameters}
}
\value{
assigns position information and PL into the environment
}
\description{
The \code{updatePortf} function goes through each symbol and calculates the PL for each period prices are available.
}
\details{
Note that the portfolio will be marked on every time stamp where prices are available.
As such, your \code{Dates} range must reflect timestamps which appear in the price stream.
Also note that you probably don't want to mark the portfolio on every tick,
}

